-removed VMA -completed UserInput -added ECS stuff -implemented fps camera controller -implemented GlobalVar class

+++ VEngine/Code/Graphics/Model.cpp
@@ -116,7 +116,7 @@ VEngine::Model::Model(const char *filepath)
 		}
 	}
 
-	indexCount = indices.size();
+	indexCount = static_cast<uint32_t>(indices.size());
 }
 
 VEngine::VKBufferData VEngine::Model::getIndexBufferData() const
